Corrections

+++ src/Service/PictureService.php
@@ -22,8 +22,12 @@ class PictureService
         return $this->fileService->processImage($imageFile, $folder, $width, $height);
     }
 
-    public function delete(Image $image, string $folder): bool
+    public function delete(?Image $image, string $folder): bool
     {
+        if($image === null) {
+            return true;
+        }
+
         // Delete image from database
         $this->em->remove($image);
         $this->em->flush();
@@ -38,27 +42,32 @@ class PictureService
         return true;
     }
 
-    public function updateProfilePicture(User $user, UploadedFile $newPicture, string $folder): bool
+    public function updateProfilePicture(User $user, string $folder): bool
     {
+        if($user->getPictureSlug() !== null){
         // remove current profile picture
         $currentPictureProfile = $this->imageRepository->findOneBy(['name' => $user->getPictureSlug()]);
 
         $delete = $this->delete($currentPictureProfile, $folder);
-        $this->setProfilePicture($user, $newPicture, $folder);
 
         if (!$delete) {
             return false;
+        } else {
+            return true;
+        }
+           
         }
 
         return true;
     }
 
-    private function setProfilePicture(User $user, UploadedFile $newPicture, string $folder): void
+    public function setProfilePicture(User $user, UploadedFile $newPicture, string $folder): void
     {
         // set the new profil picture
         $profilPicture = $newPicture;
         $profilPicture = $this->processImage($profilPicture, $folder);
 
+        $user->setPictureSlug($profilPicture->getName());
         $this->em->persist($user);
         $this->em->flush();
     }
